Move NodeId to new file
